runtime: update image hash resolution logic
The primary issue this resolves is cases when the image was loaded into
the store via `rkt fetch`, e.g. as is done with the server components in
kube-up.

Prior to this, the `rkt fetch --store-only docker://imageName` would
fail because the `docker://` prefix only works if the internal
`docker2aci` is used via fetching from a remote registry.

This change allows any image that has the appropriate "originalname"
annotation to work, regardless of source.

That does mean that the local image needs to have been converted with a
recent version of docker2aci, but that's reasonable.