rktlet/cli/init: set `--service-type=notify`
Otherwise, an error in the run command is not detected and `systemd-run`
returns 0 in any case. (`systemd-run false ; echo $?` versus
`systemd-run --service=notify false ; echo $?`)

With this change, an error in a transient unit (as described in
https://github.com/kubernetes-incubator/rktlet/issues/108) would be
easier to debug, as `RunPodSandbox` now returns the actual error to the
kubelet.