Convert cgroup-parent to a systemd slice name
When the kubelet is started with --cgroups-per-qos, it will request pods
to be created in a specific cgroup. Since rktlet starts rkt as a systemd
unit, it needs to specify a systemd slice name that represent the
requested cgroup.

Unfortunately, not all cgroup paths are representable as systemd slice
names. The Kubelet must be started with --cgroup-driver=systemd
(CGROUP_DRIVER=systemd in hack/local-up-cluster.sh), otherwise the
cgroupParent will not be convertible.

Fixes https://github.com/kubernetes-incubator/rktlet/issues/44