rktlet: add network plugin name flag
According to the CRI Networking Specification[[1]], the runtime shim is
responsible of managing pod's network life cycle. However, the current
CRI API doesn't expose options like `network-plugin`, `cni-bin-dir`, or
`cni-conf-dir`.

This adds a new `--net` option to the rktlet so users can specify what
network they want to use for pods. We rely on rkt to set up the network
so these are the paths where rkt looks for CNI configurations and
binaries:

* Network configurations: `/etc/rkt/net.d`
* Plugin binaries: `/usr/lib/rkt/plugins/net`

Note that rkt also includes some plugins[[2]] in the default stage1.

[1]: https://github.com/kubernetes/community/blob/c1eab08/contributors/devel/kubelet-cri-networking.md
[2]: https://github.com/rkt/rkt/blob/v1.28.1/Documentation/networking/overview.md#built-in-network-types