runtime: start sandboxes with --debug
Otherwise, if `rkt app add` or `rkt app start` fails, we won't get any
logs in the journal and we'll be blind to the causes of the failure.

With this, users can at least run on the host:

    # journalctl -M rkt-$POD_UUID

and see why the app failed to start.

As a complete solution, we should somehow capture this journal output
and propagate it to the error returned in StartContainer().