cmd: add --version flag
It's set at build-time to the output of `git describe --dirty --always`.

At the moment it just prints a git commit hash because there is no
annotated tag. When we add the first tag, it should start working as
intended and print versions in the format `0.1.0`,
`0.1.0-${COMMITS_SINCE_LAST_TAG}-${COMMIT_HASH}`, or
`0.1.0-${COMMITS_SINCE_LAST_TAG}-${COMMIT_HASH}-dirty`.